updating dim location staging code

+++ sql_scripts/2. staging_location_dim.sql
@@ -1,10 +1,12 @@
 --create if not exist
-DROP TABLE  computer_std.dbo.dim_location;
+IF OBJECT_ID('computer_std.dbo.dim_location', 'U') IS NOT NULL
+    DROP TABLE computer_std.dbo.dim_location;
+
 CREATE TABLE computer_std.dbo.dim_location (
     Location_ID INT IDENTITY(1,1) PRIMARY KEY,
-    Continent VARCHAR(250) NULL,
-    [Country or State] VARCHAR(250) NULL,
-    [Province or City] VARCHAR(250) NULL
+    Continent VARCHAR(250),
+    [Country or State] VARCHAR(250),
+    [Province or City] VARCHAR(250)
 );
 
 -- insert data into the table
@@ -15,4 +17,4 @@ SELECT DISTINCT
 FROM computer_std.dbo.pc_data;
 
 --3. select all the data from dim table
-SELECT * FROM computer_std.dbo.dim_location;
+SELECT * FROM computer_std.dbo.dim_location
